[lcmtypes] Refactor build file to rely less on globbing
Instead of directly using globbing for the build rules, we write out
the list of source files by hand, and use globbing to cross-check it.

In general we should not use bazel globbing for locating source files,
because it makes the build less reproducible vs local edits.  It also
turns out that soon we're going to need access to the list of sources
from another package (pydrake.lcmtypes), and sharing glob output
across package boundaries is impractical.  Writing down the list of
files makes it much easier to share and re-use.