Fix error in ContactResultsToLcmSystem
The fundamental bug is that transmogrified versions of the system were
broken -- they had an incomplete set of tables and *no* ports.

This fact was masked by the haphazard unit tests. The unit tests have
been completely revisited to provide greater coverage with special
attention paid to transmogrified systems and documenting the testing
philosophy and strategy.