[common] Deprecate DRAKE_ASSERT on a bare pointer (#15597)
For a raw pointer (e.g., 'foo' in 'void DoStuff(Foo* foo);'), when checking
for null we'd like to insist on an explicit mention of nullptr:

Correct: DRAKE_ASSERT(foo != nullptr);
Correct: DRAKE_DEMAND(foo != nullptr);
Incorrect: DRAKE_ASSERT(foo);
Incorrect: DRAKE_DEMAND(foo);

Because the macro sometimes results in error messages seen by users,
we should err on the side of extra detail about the failure condition.  The
meaning of "foo" isolated within error message text does not make it clear
that a null pointer is the proximate cause of the problem.