Switch to per-context dummy CacheEntryValue for thread safety. (#15564)
Previously we used a singleton "dummy" CacheEntryValue as a /dev/null-like destination for throw-away writes. This caused trouble with Tsan in multithreaded PR #15537.

This commit switches to using a per-Context dummy rather than a singleton. This is mildly tricky due to the need to switch dummy destinations during Context cloning.