Correct internal nomenclature for collision filtering
The public documentation refers to sets of "invariant" collision filters
(e.g., geometries fixed to a common frame are always filtered). Internally,
it was referred to as a "permanent" filter. This changes the internal
implementation to refer to them as "invariant" filters.

(This also paves the way for adding "transient" modifications to collision
filters as that uses the term "persistent". The confusion between
"permanent" and "persistent" is much higher than that with "invariant"