Enable drake visualizer to consume new hydro lcm fields
This gives the drake_visualizer plugin the power to consume the new
per-body fields in the lcm message: model and geometry names.

Previously, the _Contact class was synonymous with per-body-pair contact.
That is no longer true. A body pair can have multiple contacts. This led
to a refactoring where there is now _BodyContact (all of the contacts
between a pair of bodies) and _Contact (a single contact between body
pairs).

The names displayed in the object model for bodies and contact data now
depend on a GUI configuration. Default behavior is the legacy behavior
(compact, yet possibly ambiguous). Turning it on provides verbose but
unambiguous names all the time.

With this change, the visualizer can now handle multiple contacts per
body pair and distinguish between two bodies with the same name, but in
different model instances.

This incidentally extends simple_contact_surface_vis.cc to populate all of
the new fields in the lcm message to test the visualizer.