[geometry] Establishes Meshcat in C++ (step 1) (#15618)
This is the first of a series of PRs that will provide Meshcat as a visualizer in C++.  The design and PR strategy is documented in #13038.

This is the first PR:
Meshcat proof of life. Starts the server, demonstrates that clients can connect, and just sends one type of message to show that data can flow. Reviewers can focus on the build system and websocket server details.

The dependencies added here are all quite lightweight, and properly licensed.  And at the end of the PR train, we will likely want to deprecate meshcat-python, and eventually remove the pip dependencies that come along with it.