[geometry] Establish Meshcat in C++ (step 3)
Implements the most essential elements of the Meshcat API:
- Added SceneTree data structure.
- Added stduuid (to create the uuid's required by meshcat)
- Added SetObject from Shape, Rgba
  - Supports all shapes that meshcat_visualizer.py supports (more, in fact)
  - Does not support texture maps yet.
- Added SetTransform
- Added Delete
- Added SetProperty<double>
- Added HasPath, GetPacked* methods, primarily for testing.

This PR is > 500 lines, but a significant portion of this commit is relatively boilerplate management of the msgpack data structures used to communicate with the javascript client.

Updated the meshcat_manual_test to step the user through the checks.
Updated meschat_test to hammer on the new API.