py value: Register Value[float] for use with ProximityProperties (#15641)
Issue 15640 showed that when a float was placed into a ProximityProperties
in python, C++ code attempting to extract double failed -- it found
a type-erased python object.

This extends the values bindings to include the double typed stashed
in a Value and adapts the tests accordingly to confirm that the value
can be extracted in C++.

This introduces a unique C++ function, bound in pydrake.geometry.testing
to facilitate the testing.