[setup] Adjust removal date further into the future (#15700)
There is no great rush to tidy up the setup scripts, and if a user
happened to pause their Drake upgrades during the summer, they would
be bitten by this, with a difficult recovery.  Push the date out an
extra two months to allow for more time for the transition.