multibody plant: Fix TAMSI threading hazard (#15690)
* multibody plant: Fix TAMSI threading hazard

Closes: #15674

The storage of the TamsiSolver as member data created a thread-race hazard for
context-per-thread usage. Instead, store the solver (via shared_ptr) in a
scratch cache entry. Reorganize transmission of stiction_tolerance to be
maximally lazy, in part to avoid the absence of context access where it used to
be transmitted.

The patch also adds some minimal test coverage for context-per-thread usage in
both discrete and continuous modes.