Enable hydro geometry vis in python, implement for python meshcat
Binds the methods maybe_get_hydroelastic_mesh().

The visualizer can now take a additional arguments that will determine if
proximity geometries are visualized by the declared primitive or by any
available hydroelastic mesh representation.

It extends the arguments for parsing to include the control and modifies
show_model to benefit from the new parameters.

To see the effect:

    cd drake
    bazel build \
        //tools:drake_visualizer @meshcat_python//:meshcat-server \
        //manipulation/util:show_model

In one terminal
    ./bazel-bin/external/meshcat_python/meshcat-server

In the other terminal:

    ./bazel-bin/manipulation/util/show_model --meshcat default \
        bindings/pydrake/multibody/test/hydroelastic.sdf

You will see *nothing* - there are no visual geometries. Then try:

    ./bazel-bin/manipulation/util/show_model --meshcat default \
         bindings/pydrake/multibody/test/hydroelastic.sdf \
         --meshcat_role proximity --meshcat_hydroelastic

Now you will see a smooth, yellow sphere. Then try:

    ./bazel-bin/manipulation/util/show_model --meshcat default \
        bindings/pydrake/multibody/test/hydroelastic.sdf \
        --meshcat_role proximity --meshcat_hydroelastic

The smooth sphere will be replaced with a coarsely tesselated sphere. If
you modify the value on line 57 of the sdf file from

      <drake:mesh_resolution_hint>0.2</drake:mesh_resolution_hint>

to

      <drake:mesh_resolution_hint>0.4</drake:mesh_resolution_hint>

the sphere becomes an octahedron. Modify it to

      <drake:mesh_resolution_hint>0.1</drake:mesh_resolution_hint>

and it becomes smoother.