ConvertVolumeToSurfaceMesh moves out of internal namespace
We're putting this in part of the public API so that visualizers can use
this method to map VolumeMesh to a SurfaceMesh of the mesh boundary. This
also includes bindings and tests on those bindings.

As part of this, we move the implementation into the .cc file and
explicitly define it on the supported scalar types (with appropriate
modification of doxygen to bring it up to current practices).