[geometry:meshcat] SetCamera (#15728)
- Adds SetCamera for PerspectiveCamera and OrthographicCamera.
- Adds utility methods to allow meshcat to be used like a planar (2d) visualizer, by setting the camera to orthographic, etc.

This PR introduces a new pattern, as we work towards exposing more of the meshcat/three.js API through C++: we define the camera parameter structures in the public API, but define the message-pack packer in the private header.  I envision us using this pattern often in the future as we expose even more.