[geometry] Hydroelastic contact vis uses "minimally unique names" (#15741)
The "minimally unique name" is predicated on whether a body has a unique
name across *all* model instances. If it does, it can be displayed
simply as "body" in visualization. Otherwise, it must be
displayed as "model/body".

Furthermore, on a per-contact surface resolution, it determines if
displaying geometry names is *necessary* or not. It is necessary
if either body in contact has multiple collision geometries.

This extends the lcm message to give the visualizer sufficient information
to make the display decision.

 - MBP provides the API to report how many bodies share a name.
 - LCM message is extended to include uniqueness data.
 - Visualization removes the option to toggle "full names" and just always
   uses the minimal names.

This has *one* regrettable artifact.
 - geometry names registered by MBP are defined as
   "model name/geometry name". This comes from using MbP.GetScopedName().
   It is not clear that applying that to geometry has any value.