[geometry:meshcat] Add UI controls to Meshcat (#15768)
* [geometry:meshcat] Add UI controls to Meshcat

Adds support for buttons and sliders inside the meshcat browser.  This
is also the first instance of a bidirectional communication (the new
direction being browser => c++).  I've been using it for class for a
week or so, and it's been working well under that stress test.

Breaking change: Fixed a spelling error in the MeshcatVisualizerParams
field:
delete_on_intiliazation_event => delete_on_initialization_event.