Clean up Drake-owned URDF files: remove non-spec tags and attributes
In the past, we've had various custom tags and attributes that Drake
consumed. Some of them are no longer consumed. It includes:

  - collision@group
  - visual@group
  - collision/drake_compliance/* (now relplaced by
    collision/drake:proximity_properties/*
  - joint/dynamics@coulomb_window

Presence of some of the ancient tags cause warnings to get written to
the console. There's no reason for these tags/attributes to be present
in anything that Drake owns. Where the tags/attributes are simply dead,
they've been removed. Where they've been replaced, the tags have been
updated.

This also updates the comments for two Atlas URDFs, removing the
out-of-date warning that the files were generated from xacro files and
should not be edited.