Remove final vestiges of mesh feature index types
There are four index types associated with the two mesh types. This final
commit removes the place-holder aliases for those index types, completing
the removal. The aliases served the role of deferring any action that had
to be coordinated based on the removal of multiple index types.

This also does the work that had to be deferred until all index types
could be removed completely:

  - Constructs that had to work across mesh types, had to worry about
    specific index types associated with mesh types. They no longer have to
    worry about those details and can simply use ints. This includes:
      Bvh, Aabb, Obb, MeshDeformer, MeshFieldLinear, and writing meshes to
      VTK.