For Identifier::get_new_id() to print out the id flavor
To aid in illustrating a bug, this changes get_new_id() to report which
identifier is generating an id. In order to use NiceTypeName to this end
the various Identifier<class FooTag> invocation required a tweak; the
FooTag needed to be a real class for the type system to work. So, this
also touches *all* incarnations of Identifier<class FooTag> to back up
FooTag with an actual definition.