Replace LDLT in multibody plant with math::LinearSolver
The math::LinearSolver is a lot faster than templated Eigen solver when
the scalar type is AutoDiffScalar.
Add a default constructor for LinearSolver class.
Change the return type of LinearSolver::Solve(). If both A and b contain
the same scalar type (either double or symbolic::Expression), then the
return type is Eigen::Solve<>, the same return type as
decomposition.solve().