Replace history-deprived files with alternatives
In PR 15859, an accident killed the history of pydrake/geometry_py.cc
and pydrake/test/geometry_test.py. The previous merge contortions were
to create pydrake/geometry_py_all.cc and pydrake/test/geometry_all_test.py
which carried that history.

This last commit finishes the work of replacing the old files with the new
in the build system. Moving forward, the new files carry inheritable
history that can contribute to the decomposition.

Incidentally, this also restores the changes from #15856. The superfluous
FakeId in the bindings had been restored during the breakout and this
restores it to the post #15856 state.