Make SurfaceVertex and VolumeVertex alias for Vector3
This allows me to remove all of the invocaitons of r_MV() without also
having to change signatures.

The sole loss in using the alias is that we no longer automat type
deduction from, e.g.,

    double foo = 10;
    SurfaceVertex(foo, foo, foo);

In order to avoid creating python aliases for C++ aliases, I've simply
removed the python bindings entirely.