Removing VolumeVertex and SurfaceVertex (#15903)
This replaces SurfaceVertex and VolumeVertex with Vector3. In some
cases it leads to a code restructuring. Where code would iterate
through the `VolumeVertex v` and then, inside the loop, create the
*actual* loop variable `p_MV`, we've changed the for loop to
simply iterate through `p_MV` directly.

This touches all usages and removes the ubiquitous invocations of r_MV();
it is no longer necessary or valid, the vertex *is* its own position.