[hydro] Remove SurfacFaceIndex type (#15932)
There are four index types associated with the two mesh types. This removes
the third of the four indices.

In this case "removal" means:
  1. Removing all usages of SurfaceFaceIndex in Drake C++ code.
  2. Removing the python bindings.
  3. Keep the type, but change it to be an alias to an int. This will allow
     us to defer updating Anzu until *all* index types have been removed.
     At that point, we'll remove the int aliases and update Anzu.
     - TypeSafeIndex allows for index vs unsigned int comparisions. This
       alias will break this functionality. Where necessary, those signed-
       unsigned comparisons are also updated.

Note: There are three instances where the removal of the index type
represents, in some sense, a loss:

  - mesh_intersection.cc, IsFaceNormalAlongPressureGradient()
  - surface_mesh.h: SurfaceMesh::CalcGradBarycentric
  - bvh.h: BvttCallback

However, the loss is considered acceptable. In the first two cases, the
API is internal or private, we can rely on developer discipline and unit
tests to make sure that the internal code is using the correct
interpretation of index types.

In BvttCallback, there would *ostensibly* be value in distinguishing the
index types of the two meshes being collided. However, that distinction
only exists if the mesh types themselves are different. If they are the
same (e.g., for compliant-compliant contact), the types provide no
distinguishing power. So, the code has to be properly tested for correct
ordering, regardless of types. So, the constraint of the typed indices
provides more cost than benefit.