Parse frames with model instance when adding bushings/joints
Enforces bushing and joint declarations to use frames that have been
declared in the same urdf/sdf file. It does this by now requiring the
frame name to be part of the current (the one being parsed) model
instance. Previously, parsing one file could inadvertently reference a
frame added by a previous file. Any workflow that relied on that
implicit cross-file integration will now break. On the plus side, it
allows a model which declares a bushing to be loaded multiple times.