[geometry] Add MeshcatCone isa Shape (#15936)
* [geometry] Add MeshcatCone isa Shape

This supports visualizing cone geometries in meshcat (which is
immediately useful for drawing the arrowheads of contact force
vectors).

Per a discussion with @SeanCurtis-TRI, we've decided to add the type
to the main Shape class hierarchy, but with a class name
(`MeshcatCone`) that makes it obvious that the current support for
this shape is limited to Meshcat.