Convert SuiteSparse "patch" to a real patch
Replace the modified SuiteSparse_config.mk that was being injected into
the SuiteSparse source tree with an actual patch. The reason we weren't
using a patch before is because the changes originate from Fletch¹,
which was used (after much simplification) as the basis for building our
dependencies. Unlike us, Fletch is not able to assume that git is
available, and so is unable to apply real patches and must resort to
having copies of files that need to be modified. We, of course, don't
have that limitation, and a proper patch should be easier to maintain
and have less question as to distribution rights.

(¹ https://github.com/Kitware/fletch)