Optimize wheel tests
Refactor the wheel test driver to build provisioned containers for
testing the wheels (that have the correct Python already installed).
This allows us to skip some common steps in order to make incremental
testing faster.

Note that this does not affect how either of the scripts need to be
invoked.