Enable snopt when building wheels
Modify wheel build scripts to enable snopt, using ssh agent forwarding
from the build host for authentication. Modify provisioning to enable
ssh and to permit accessing github.

Note that this requires BuildKit or a recent version of Podman. It also
requires that the builds are run in an environment with an active ssh
agent which is able to fetch snopt from Github without additional
configuration (i.e. must work without .ssh/config).