[wheel] Build current checkout, not a fresh clone
Modify build script to create an archive of the current repository
checkout, and to use this rather than cloning the latest master. This
will make some testing easier, as well as enable CI to build wheels
against pull requests.

Also, avoid use of readlink in test-wheels.sh in preparation for adding
macOS support.

Co-authored-by: Jeremy Nimmer <jeremy.nimmer@tri.global>