Rewrite pip build script in Python
Replace the build-wheels.sh script with one written in Python. This
allows for much greater flexibility in what the script does (which is
useful for testing and debugging) without the need to edit the script to
e.g. select what wheel(s) to build. It also provides a small reduction
in duplicated code when adding additional target platforms.

One major change is that the new script uses `git archive` to locally
generate a tarball of the Drake repository, which is then fed into the
Docker container to be built, rather than performing a `git clone`
inside the container. Particularly, this means that the container(s)
will build whatever tree the user has checked out, rather than 'master'.
This is particularly a requirement to enable CI builds on pull requests,
but should also make testing local changes easier.