[pydrake] Port values borrowed from a Context keep the Context alive
In a Jupyter notebook, if a helper function creates a local context
and then returns a value computed by an abstract-valued output port
evaluated against that context, and if the return value is a C++ class
type that is bound in pydrake, and if the function does not retain or
return the context itself such that the context is garbage collected,
then the storage that backs the return value would be erroneously
deleted at the conclusion of the cell, thus leaving the return value
pointing at garbage memory.

This commit adds a reference_internal annotation on the value into the
context, so that the context will not be destroyed in this case.

There does not seem to be any way to trigger these conditions outside
of a Jupyter notebook, so the regression test here takes the form of
an ipynb file.