Use working tree to build wheels
Rewrite the logic to generate the source tarball for building wheels to
copy the local tree, rather than using `git archive`. This allows local
changes to be tested (but introduces the risk of forgetting to commit
such changes). As a minor side benefit, this removes the use of the git
Python bindings.