Eliminate representative triangle contact surface
This simply eliminates the tentative API for producing a contact surface
with fewer elements. It is being removed so that it doesn't interfere
with introducing the polygon surface mesh version of the contact surface.

This means, there will temporarily be no "low resolution" contact surface
in discrete mode until the PR train outlined in 15796 is completed.