[hydroelastic] Modify contact surface visualization LCM (#16102)
This changes the representation of a mesh in the contact surface lcm
message. This change makes it forward compatible with when we want to
render polygonal meshes. It leads to the removal of a custom triangle
message for contact surfaces.

The various call sites that are responsible for generating such a message
have been updated to write to the new format. Incidentally, various
variable names (e.g., msg) that were *not* styleguide compliant have been
updated to remove the bad abbreviation.

The drake_visualizer plugin has been likewise modified to handle the new
representation. In addition, how it handles drawing edges on the contact
surface has been streamlined (it uses VTK functionality for drawing
hidden line objects instead of creating multiple meshes).