[meshcat] Avoid dropping messages due to backpressure, and provide Flush()
Resolves #16109

After reading through the uwebsockets code, I understand now that the
maxBackpressure websocket option was *only* being used to decide when
to drop packets.  We never want to drop packets, so setting that to
zero is the correct behavior.

Backpressure is still a legitmate concern.  If we keep shoving data
into the send buffer faster than the websocket can send, then
backpressure will build up (consuming memory and causing the
visualizer to fall behind the main thread).  I've provided a `Flush()`
command for Meshcat which allows users to manually sync the main
thread if it is useful/needed in their application.