Get wheel version from build-wheels
Modify the wheel building logic to expect the wheel version to be passed
in via an argument to the Dockerfile, and add the wheel version as a
required argument to the build-wheels script.

This is a somewhat lateral change with respect to the previous behavior
of hard-coding the version in the setup.py; however, allowing the
version number to be handed down from e.g. a CI job will simplify other
tasks, such as subsequently testing wheels, since we lack a good
mechanism to extract the version number from a build.