Provides MeshPainterSystem + example
This code is pulled almost directly from Anzu; the primary author who
deserves all credit is SeanCurtis-TRI.  A number of MIT 6.843 students
wanted this functionality for their final projects.  I do not intend
to merge it; at least immediately.

The changes to the example source files are mostly changes to the
namespace and the resource paths.  I've integrated it into the drake
build system and added python bindings.  I did have to produce a new
render_engine factory, and do some minor surgery to the source code to
avoid installing the vtk header files.  And fixup the documentation a
bit to make mkdoc succeed.

There are also unit tests in Anzu which should be ported if we ever
want to do more with this.