Enable polygon contact surface in mesh-half space intersection
This reframes mesh-half space intersection to use the new MeshBuilder so
it can create either polygon or triangle mesh representations. The APIs
have been modified to communicate the desired representation and the tests
have been updated accordingly.

This entailed changes to the organization of the code so it more closely
resembled mesh-plane intersection.

Various related call sites have been updated to *explicitly* request
triangle meshes (this was implicit before).