[hydro] Correct bug in MeshFieldLinear::TransformGradients() (#16160)
The field is expressed in a frame M. When transforming it to a frame N,
we call `Transform()` (formerly known as `TransformGradients()`). If the
field has stored gradients, this re-expresses the gradients. However,
that is insufficient. EvaluateCartesian() uses those gradients to evaluate
the field (when they exist). But it also depends on `values_at_Mo_` --
another frame-dependent quantity.

This updates the function in question and expands the unit test to
cover this previously uncovered aspect of the code.