[hydro] Add MeshBuilder -- interface for building ContactSurface meshes (#16134)
The MeshBuilder classes extract the knowledge of how to add intersection
polygons to mesh representations. The intersection algorithms can simply
generate the polygon (using whatever logic they require) and pass the data
over to the builder.

This also renames the internal `AddPolygonToMeshData()` function to better
reflect its purpose and contrast it with the new
`AddPolygonToPolygonMeshData()` function.