Create initial proof-of-concept render server
Create a simple REST render server using flask and f3d (a small VTK
application for viewing glTF). This, especially the requirement for f3d,
is intended to be an early proof-of-concept. The server should be usable
with the intended render client, but also provides some very basic web
pages which allow it to be used via a web browser.