Adds suffix strings to give unique names to each position/velocity el… (#14971)
* [multibody] Provide suffix strings for each position/velocity in MultibodyPlant

This PR plumbs through the mobilizer/joint class hierachy in order to
accomplish a simple goal: I want to have string names for each element
of the multibody context.  We already have names for bodies and joint,
but (floating) bodies and joints can have more than one
position/velocity.

Here we introduce a "suffix" for each position/velocity element which
can be combined with the body/joint name to obtain a meaningful
position / velocity name.