Create initial proof-of-concept render server
Create a simple REST render server using flask and an external program
for producing the rendered image. (Currently, this is written to assume
f3d; a small, VTK-based glTF viewer). The server should be usable with
the intended render client, but also provides some very basic web pages
which allow it to be used via a web browser. Use of an external program
to produce the actual rendered images will allow for different renderers
to be easily swapped in and out.