[install] Refactor install.py.in toward testability
Relevant to: #7331

This patch extracts most of the python code from install.py.in to a new
module install_lib.py, and adjusts bazel rules and targets to work with
the new layout.

No actual new tests are added at this step.

The moved code in the new file is only adjusted so that python lint
checks will pass. The new function `execute_installation()` is
essentially the old `main()` function, with an extra callable argument.

At this stage, all of the existing tests should pass and existing
install rules should still work. The next step toward closing #7331
would be to increase test coverage of the install_lib.py code.