Refactor drake-visualizer build
Overhaul the Docker machinery for building drake-visualizer. This
reworks everything to use a series of scripts and supplemental files
(e.g. CMake arguments and the list of distro packages to install are now
separate) rather than trying to mash everything directly into the
Dockerfile. The new stuff is substantially modeled after the Docker
machinery used to build the Drake wheels.

Additionally, since we are moving to Director needing a different
version of VTK than is used by Drake itself, the new machinery builds
VTK as part of the overall process, rather than relying on an already
existing package.