"Clean up" wheel scripts
Add slightly more strict error checking to the various Wheel scripts.
It's unlikely this will actually catch anything "new", but it's mostly
being done to increase consistency across various scripts of this
nature, particularly with the (still in progress, as of writing)
refactoring work being done to the VTK scripts.