[trajectory_optimization] MultipleShooting has-a MathematicalProgram (#16333)
BREAKING CHANGE.  This PR bends over backwards to deprecate things
properly, but there could still be cases for which this is a breaking
change.  For instance, if user code creates a cost/constraint class
that takes a MathematicalProgram as an argument, it was previously
possible to pass a MultipleShooting instance to that method directly.
Now you must pass MultipleShooting.prog().

Previously the trajectory optimization base class MultipleShooting
derived from MathematicalProgram (MultipleShooting is-a
MathematicalProgram).  This PR changes that relationship to
MultipleShooting has-a MathematicalProgram (as a member variable).

This is the pattern used in much of our other optimization classes,
such as InverseKinematics and the Anzu
KinematicTrajectoryOptimization.  It has a number of advantages,
including the ability, for instance, create multiple trajectory
optimization problems in the same mathematical program.