[tools] Fix "generated by these conflicting actions" in Bazel 5.0
When something depends on a unit test (in our case, the macOS dSYM
generation), then Bazel as of 5.0 dies a horrible death.

Work around that by adding the dependency only in cases where the user
has opted-in to debugging on macOS, and in that case also turn off the
problematic trim_test_configuration feature.