[multibody] Guard hazardous map access by GeometryId
The GeometryId values for collisions come in on an input port.
We need to sanity-check them prior to using them.

For two maps that were already dead code, we just kill the deadwood
instead of adding guards for it.

The related default_coulomb_friction_ was also dead code, so is
also purged.