[multibody/parsing] Support default geometry attributes in MuJoCo parser
Towards #16369

Implements the default settings as described in https://mujoco.readthedocs.io/en/latest/modeling.html#cdefault

I had originally planned to not have to support default in my initial
parser implementation.  But the DeepMind control suite (which we
switched to using for the unit tests) uses default heavily, and the
files cannot be parsed correctly without this support.