[meshcat] Remove accidental support for Expression
The documentation and template definitions said "nonsymbolic only",
but the scalar conversion still accidentally allowed for Expression.
Add the opt-out for Expression to the header.

Relatedly, the pydrake bindings had an if-check for Expression, but it
was dead code (due to the caller's list of types) so is also removed here.