[parsing] Fix mis-parse of collision filter groups
Closes: #16310

Previously, a typo'd element name caused only the first element of
`drake:ignored_collision_filter_group` within a group to be actually
processed.  This patch fixes the typo and extends the existing tests to
cover the error case. Regrettably, it does not attempt a more vigorous
test refactoring. That may happen in the near future.