[common] Give RandomGenerator efficient default and move constructors
Copying 5000 bytes of storage during a move is wasteful; better to just
swap a heap pointer.

Calculating 624 random integers in the default constructor is wasteful,
if we're only trying to allocate a storage for generator that will be
moved into later (e.g., when resizing a collection of generators).

While we're here, tidy up some of its API documentation for completeness.