[solvers] Allow evaluators to provide gradients w/o AutoDiffXd
Consider the cost of using EvaluatorBase for a Cost/Constraint with a
huge number of parameters (like the ones we get with
MultilayerPerceptrons).  On each call to Eval(AutoDiff), we using
InitializeAutoDiffXd(x) to create an enormous set of .derivatives() to
inefficiently represent the identity matrix, which the Evaluator
doesn't even want to use.  Furthermore, because the method *could* be
called with any AutoDiffXd on the input, each implementation must
implement the chain rule to compute dydx*dxd__ , when we really just
want dydx.

This PR provides a shortcut, so that implementations of EvaluatorBase
can provide the relevant gradients (only).  In the short term, derived
classes will also have to implement the DoEval(AutoDiff..), or throw.

I've updated only snopt_solver to use this so far, just to initiate
the concept.