[geometry] Factor our uWS callbacks for readability
This is also in anticipation of wrapping all callbacks with exception
handlers, which will be easier if they have real names.

Add an implementation comment to all of the PIMPL functions clarifying
their scope, now that we have a wider variety of happenings.

Fix an indentation-level snafu on some of the PIMPL functions.