[wheel] Add missing apt-get update
Because the Dockerfile will cache based on the contents of these files only,
and not the state of the internet hosts, we run the risk of security versions
disappearing between the cached provision-base step and a new provision-python,
resulting in the inability to build.

We must maintain the invariant that any shell script with 'apt-get install'
must always do 'apt-get update' first.