[pydrake] Ensure GIL is locked when logging from a C++ worker thread
To avoid lock order inversion, we must avoid using any mutexes in the
spdlog code paths; instead we should grab the GIL once we've reached the
Python sink.

This means that users will not be able to log to both a C++ sink and
Python sink anymore. Users who want to do so can configure logging on
their own, in which case Drake's logging reconfiguration will leave it
unchanged.