[geometry] Fix Meshcat crash with duplicated controls names
When asked to add a control with a duplicate name, throw an exception
instead of trying to implicitly delete the existing control.

Deleting the existing control never worked anyway; it was checking for
duplicates using the moved-from name, which was always empty. This led
to duplicate names in the controls_ vector, and then segfaults.

Add more internal assertions that the controls_ vector is congruent
with the buttons_ and sliders_ vectors.