Simplify several third-party inclusions
Refactor how we import several third-party packages to follow the
pattern (already in use elsewhere) where we externally declare the
library directory, such that we can share the build rules across
platforms. This reduces the total number of files that need to be
maintained.

For the sake of simplicity and consistency, also use -rpath in more
places. (Some of this happens as a consequence of the simplification, as
macOS needs it explicitly while it is potentially superfluous on
Ubuntu.)

Last, fix some manylinux imports which were pulling system versions
rather than the versions built as part of the wheel build.