[parsing] Cleanup detail_urdf_geometry
Relevant to: #16229

Move the anonymous namespace to the top of the cc file, as is
customary. Move a doomed legacy function into the anonymous namespace;
since it is not declared in the header, it should not leak its symbol at
link time.

This patch should have 0 functional change, other than hiding the stray
link symbol.