[parsing] Introduce class into urdf parser implementation
Relevant to: #16782

This is the first part of plumbing error channels into the URDF
parser. Since the whole change set will have all the grace of a snake
eating an elephant, this step is limited to introducing some
infrastructure and converting one function.

Subsequent patches will convert the rest of this module and plumb error
channels to other modules.