[parsing] Port remainder of detail_urdf_parser to diagnostics
Relevant to: #16782

This patch completes the conversion of detail_urdf_parser to use the
passed-in diagnostic policy for reporting warnings and errors. Some
parsing routines in related modules are not yet ported.

The patch boosts test coverage of error and warning branches to 100%, by
adding numerous (25-ish?) new test cases. It also converts test cases
from relying on throw-and-catch mechanics to message capture by a
test-only diagnostics policy.

As collateral damage, the collision filter parsing routines in
detail_common gain some ability to tolerate parsing functors that return
on error, instead of throwing.