Minimize libtiff build (for wheels)
When building libtiff (for wheels), explicitly disable some things that
we don't need. This should have no effect for the manylinux builds, as
those are fully hermetic and don't have anything we didn't ask to have
installed, but will be very important on macOS (especially for local
testing) to ensure we don't pick up libraries that happen to be
installed from brew.