[pydrake] Stub out scipy.sparse for unit testing (#16840)
In order to reduce our install footprint, pydrake will no longer depend
on scipy on or after 2022-07-01.

Any users of the few scipy.sparse-returning methods in pydrake will need
to add a direct dependency onto scipy to their own projects. The methods
that return a sparse matrix will raise an ImportError when they are
called when scipy is not installed.